pyproject.toml
[project]
name = "remedy"
version = "0.1.0"
description = "AI-augmented SAST+SCA with auto-fix PRs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic~=2.9",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "redis>=5.0",
  "rq>=1.16",
  "requests>=2.31",
  "typer>=0.12",
  "langchain>=0.2",
  "google-generativeai>=0.7",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "gitpython>=3.1",
  "PyJWT>=2.9"
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
  "httpx>=0.27",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
log_level = "info"
